requirements.txt
trio-websocket
urllib3
wmi
socks
httpx
psutil
requests
keyboard
cloudscraper
requests-toolbelt
undetected-chromedriver
discord-webhook
pycryptodome
pypiwin32
browser-history
cryptography
scapy
pyasn1
pynput
nylas
pillow
pyotp
flask
colorama
cookies_package
ipinfo
folium
pystyle
Flask-SocketIO
pyinstaller